requirements.txt
Django~=4.2.5
djangorestframework~=3.14.0
djangorestframework-stubs==3.14.2
djangorestframework-simplejwt==5.3.0
pydantic==1.10.12
